hi2c1 is replaced by _i2cHandler
The I2C handler hi2c1 was originally implemented in main.h/.c in the early stages of library development. It is replaced by _i2cHandler. This variable is defined in the library SerLCD.h/.c and must be set with the function InitDisplay(). From then on any function of the library can be used.